create convenience wrapper for clearing state
